Ported load menus to the loadmenu framework
